Common trunk upstream

+++ right/src/usb_report_updater.c
@@ -48,7 +48,7 @@ bool SuppressMods = false;
 bool SuppressKeys = false;
 sticky_strategy_t StickyModifierStrategy = Stick_Smart;
 
-uint8_t KeystrokeDelay = 0;
+uint16_t KeystrokeDelay = 0;
 
 key_state_t* EmergencyKey = NULL;
